Selectors, links, img typo fix CSS, XPath, new file. Games, Android Java learning app (and ERB) links.
Also fixed typo in file extension that caused one image with test types
NOT to be displayed.

There are two ways of constructing a software design.  One way is to make it so simple that there are obviously no deficiencies. And the other way is to make it so complicated that there are no obvious deficiencies.
    -- C.A.R. Hoare